Append context fields to sip.Request/sip.Response

+++ util/rand.go
@@ -27,7 +27,7 @@ func RandString(n int) string {
 	// fill output
 	for pos := range output {
 		// get random item
-		random := uint8(randomness[pos])
+		random := randomness[pos]
 		// random % 64
 		randomPos := random % uint8(l)
 		// put into output
